packages/voice_agent/pyproject.toml
[project]
name = "voice-agent"
version = "0.1.0"
description = "LiveKit-based voice agent for AI Lead Qualification System"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "livekit-agents[silero,turn-detector]~=1.3",
    "livekit-plugins-noise-cancellation~=0.2",
    "livekit-plugins-deepgram~=1.0",
    "livekit-plugins-openai~=1.0",
    "livekit-plugins-elevenlabs~=1.0",
    "python-dotenv",
    "shared",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/voice_agent"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "A", "C4", "UP", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
